% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_evaluate_cart.R
\name{create.eval.cart}
\alias{create.eval.cart}
\title{Build & Evaluate Classificaion & Regression Trees (CART)}
\usage{
create.eval.cart(x, response, seed = 42, splitRatio = 0.7,
  minbucket = 10, cp = 0, prune = FALSE, plot = TRUE)
}
\arguments{
\item{x}{Data Frame that you want to work with}

\item{response}{Dependent(Target) Variable}

\item{seed}{Seed Value, default = 42}

\item{splitRatio}{Ratio to split the dataframe into Train and Test}

\item{minbucket}{Minimum number of observations that should be present in each leaf/terminal node}

\item{cp}{Cost Complexity Paramter Value}

\item{prune}{Pruning Tree on the basis of minimum Cross Validation Error (xerror), default = FALSE}

\item{plot}{Plot the model Tree and CP Table}
}
\value{
A List with Model Attributes & Evaluation Results
}
\description{
This function takes a dataframe as an input along with the name of the response variable.
It then sets the seed as per user or default value and split the data into training and testing sets.
After the split it builds a CART model on training data and evaluates the performance of the model on different Evaluation methods.
It can also optimize the CART tree by pruning it as per the least Cross Validation error, user has to define if they want the model to be pruned or not.
}
\examples{
create.eval.cart <- function(x, response, seed = 42, splitRatio = 0.7, minbucket = 10, cp = 0, prune = FALSE, plot = TRUE)
\dontrun{ }
\donttest{ }
}
\author{
Gupta, Deepak <deepak@analyticsacedemy.ml>
}
